test: move a listening_addr test from lib to tests::listening_addrs
